Homework/Assignment 3/Problem_1/requirements.txt
newsapi-python
kafka-python
# Add other dependencies below as needed
pandas
matplotlib
# For Spark Streaming part, you'll need pyspark
pyspark>=3.1.0 # Specify a version, e.g., 3.1.0 or higher
# For NER, you might need spacy or nltk
spacy>=3.0.0
# nltk>=3.5
# spaCy language model (install separately)
# python -m spacy download en_core_web_sm
